changed from float to double

+++ src/shapes/PlanarShape.java
@@ -12,8 +12,8 @@ import data.CompareShape;
  */
 public abstract class PlanarShape implements CompareShape<PlanarShape> {
     public abstract String toString();
-    public abstract float area();
-    public abstract float originDistance();
+    public abstract double area();
+    public abstract double originDistance();
 
     public abstract void setSize(int size);
     public abstract int getSize();
@@ -27,8 +27,8 @@ public abstract class PlanarShape implements CompareShape<PlanarShape> {
      */
     @Override
     public boolean compareTo(PlanarShape shape) {
-        float min = this.area() - this.area()*0.0005f; //0.05%
-        float max = this.area() + this.area()*0.0005f;
+        double min = this.area() - this.area()*0.0005f; //0.05%
+        double max = this.area() + this.area()*0.0005f;
 
         if(min <= shape.area() && shape.area() <= max) {
             if(shape.originDistance() < this.originDistance()) {
